Add new functionality
Added a way to know the winner, create an array with all the possible winning options, then inside a function use a for loop to do the iterations. A way was also created to check if the game has ended, this was done using the state and using the every method.